refactor(#842): Web UI→hooks→clients rule + chat de-dual-home
Closes #842
Parent #829

Written by opencode (opencode-go/deepseek-v4-flash, variant max)
