+++ apps/web/src/features/chat/components/DataPartCard.tsx
@@ -1,7 +1,7 @@
 import type { ChatDataPart } from "@animichi/contract";
 import type { ReactNode } from "react";
-import { classifyFailure } from "../../../lib/chat/error-classifier";
-import type { ChatErrorState } from "../../../lib/chat/error-classifier";
+import { classifyFailure } from "../lib/error-classifier";
+import type { ChatErrorState } from "../lib/error-classifier";
 import { isIntentOnly, parseChatDataPart } from "../data-parts";
 import type { ChatDict } from "../i18n";
 import { intentRegistry } from "../registry";
