+++ apps/web/src/features/chat/components/MessageList.tsx
@@ -1,6 +1,6 @@
 import type { ChatStatus, UIMessage } from "ai";
-import { isBypassTurn } from "../../../lib/chat/selected-points-bypass";
-import { routeDocumentKey, supersededFlags } from "../../../lib/chat/supersession";
+import { isBypassTurn } from "../lib/selected-points-bypass";
+import { routeDocumentKey, supersededFlags } from "../lib/supersession";
 import { HIDDEN_TOOL_STEPS } from "../i18n";
 import type { ChatDict } from "../i18n";
 import { formatElapsed } from "../telemetry";
