+++ apps/web/src/features/chat/components/SelectionTray.tsx
@@ -1,5 +1,5 @@
 import { useSpotSelection } from "../selection/use-spot-selection";
-import { sameIds } from "../../../lib/chat/selected-points-bypass";
+import { sameIds } from "../lib/selected-points-bypass";
 import type { ChatDict } from "../i18n";
 
 /** The recompute turn's lifecycle as the tray sees it (issue #273 S1.7 E2). */
