+++ apps/web/src/lib/route-detail/pin-state.ts
@@ -1,3 +1,4 @@
+// TODO(refactor-skeleton): de-dual-home route-detail lib -> features/route-detail/ — see #842
 import type { TimedStop } from "@animichi/contract";
 import { completedTotals, isStopCheckedIn } from "./data-state";
 import type { RouteDetail } from "./data-state";
