+++ apps/web/tests/unit/chat/basemap-mount-failure.test.tsx
@@ -11,8 +11,8 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import { attachBasemap } from "../../../src/features/bubble-map/bubble-map-controller";
 import { SearchResult } from "../../../src/features/chat/components/SearchResult";
 import { chatDictFor } from "../../../src/features/chat/i18n";
-import { toSearchSpots } from "../../../src/lib/chat/spot-clusters";
-import type { SpotRowLike } from "../../../src/lib/chat/spot-clusters";
+import { toSearchSpots } from "../../../src/features/chat/lib/spot-clusters";
+import type { SpotRowLike } from "../../../src/features/chat/lib/spot-clusters";
 
 vi.mock("pmtiles", () => ({ Protocol: class { readonly tile = () => undefined; } }));
