+++ apps/web/tests/unit/chat/save-wall-dismissal.test.tsx
@@ -13,7 +13,7 @@ import { LoginModal } from "../../../src/components/auth/LoginModal";
 import { TimedItinerary } from "../../../src/features/chat/components/TimedItinerary";
 import { chatDictFor } from "../../../src/features/chat/i18n";
 import { DEFERRED_SAVE_KEY } from "../../../src/features/chat/save/deferred-save";
-import { itineraryView } from "../../../src/lib/chat/itinerary";
+import { itineraryView } from "../../../src/features/chat/lib/itinerary";
 import { renderWithLocale, setLanguages } from "../_i18n";
 import { ujiItinerary } from "./_route-fixtures";
