+++ apps/web/tests/unit/chat/search-copy-i18n.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 import { chatDictFor } from "../../../src/features/chat/i18n";
 import { clusterName, episodeTag, spotCountBadge } from "../../../src/features/chat/search-copy";
-import type { LocatedSpot, SpotCluster } from "../../../src/lib/chat/spot-clusters";
+import type { LocatedSpot, SpotCluster } from "../../../src/features/chat/lib/spot-clusters";
 import { LOCALES } from "../../../src/i18n/locales";
 
 const CENTER = { lat: 34.89, lng: 135.8 };
